feat: migrating pointcloud types (#6996)
* feat: changed most of sensing to the new type


* chore: started applying changes to the perception stack


* feat: confirmed operation until centerpoint


* feat: reverted to the original implementation of pointpainting


* chore: forgot to push a header


* feat: also implemented the changes for the subsample filters that were out of scope before


* fix: some point type changes were missing from the latest merge from main


* chore: removed unused code, added comments, and brought back a removed publish


* chore: replaced pointcloud_raw for pointcloud_raw_ex to avoid extra processing time in the drivers


* feat: added memory layout checks


* chore: updated documentation regarding the point types


* chore: added hyperlinks to the point definitions. will be valid only once the PR is merged


* fix: fixed compilation due to moving the utilities files to the base library


* chore: separated the utilities functions due to a dependency issue

* chore: forgot that perception also uses the filter class


* feature: adapted the undistortion tests to the new point type


---------

+++ perception/raindrop_cluster_filter/src/low_intensity_cluster_filter_node.cpp
@@ -125,9 +125,9 @@ bool LowIntensityClusterFilter::isValidatedCluster(const sensor_msgs::msg::Point
     RCLCPP_WARN(get_logger(), "Invalid point cloud data. point_step is less than 16.");
     return true;
   }
-  for (sensor_msgs::PointCloud2ConstIterator<float> iter(cluster, "intensity"); iter != iter.end();
-       ++iter) {
-    mean_intensity += *iter;
+  for (sensor_msgs::PointCloud2ConstIterator<uint8_t> iter(cluster, "intensity");
+       iter != iter.end(); ++iter) {
+    mean_intensity += static_cast<float>(*iter);
   }
   const size_t num_points = cluster.width * cluster.height;
   mean_intensity /= static_cast<double>(num_points);
